+++ src/view/MenuApplication.js
@@ -7,6 +7,8 @@ import MenuAppShell              from './MenuAppShell.svelte';
 import AppStateClientSettingApp  from './app-state/client-setting/AppStateClientSettingApp.js';
 import AppStateSessionApp        from './app-state/session-storage/AppStateSessionApp.js';
 import ChatDialogContent         from './chatmessage/ChatDialogContent.svelte';
+import BasicDocumentApp          from './document/basic/BasicDocumentApp.js';
+import EmbeddedDocApplication    from "./document/embedded-collection/EmbeddedDocApplication.js";
 import HeaderButtonsApplication  from './header-buttons/HeaderButtonsApplication.js';
 import HelloFoundryApplication   from './hello/HelloFoundryApplication.js';
 import PositionApplication       from './position/app-control/PositionApplication.js';
@@ -35,7 +37,7 @@ export default class MenuApplication extends SvelteApplication
          resizable: false,
          minimizable: true,
          popOut: false,
-         width: 200,
+         width: 225,
          height: 'auto',
          positionOrtho: false,
          transformOrigin: null,
@@ -50,6 +52,8 @@ export default class MenuApplication extends SvelteApplication
                buttons: [
                   { title: 'Hello Foundry', class: HelloFoundryApplication },
                   { title: 'Header Buttons', class: HeaderButtonsApplication },
+                  { title: 'Reactive Document (Basic)', class: BasicDocumentApp },
+                  { title: 'Reactive Embedded Collections', class: EmbeddedDocApplication },
                   { title: 'Chat Message', onclick: () => new TJSDialog(
                      {
                         title: 'Essential Svelte (ESM) - Chat Message',